+++ nym-node-status-api/nym-node-status-api/src/metrics_scraper/error.rs
